+++ src/main/java/lotto/model/lottoInfo/BonusNumber.java
@@ -1,4 +1,4 @@
-package lotto.model;
+package lotto.model.lottoInfo;
 
 public class BonusNumber {
     private Integer bonusNumber;
