+++ src/main/java/lotto/model/lottoInfo/LottoInfo.java
@@ -1,4 +1,4 @@
-package lotto.model;
+package lotto.model.lottoInfo;
 
 public class LottoInfo {
     private WinningNumber winningNumbers;
