+++ src/main/java/lotto/model/lottoInfo/LottoPrice.java
@@ -1,4 +1,4 @@
-package lotto.model;
+package lotto.model.lottoInfo;
 
 public abstract class LottoPrice {
     protected Integer price;
