+++ src/main/java/lotto/model/lottoInfo/PriceData.java
@@ -1,7 +1,9 @@
-package lotto.model;
+package lotto.model.lottoInfo;
 
 import java.util.HashMap;
 import java.util.Map;
+import lotto.model.enums.Price;
+import lotto.model.enums.Rank;
 
 public abstract class PriceData {
     protected Map<Rank, Price> prizeByRank = new HashMap<>();
