+++ src/main/java/lotto/model/lottoInfo/StandardLottoPrice.java
@@ -1,4 +1,4 @@
-package lotto.model;
+package lotto.model.lottoInfo;
 
 public class StandardLottoPrice extends LottoPrice {
     private static final Integer STANDARD_LOTTO_PRICE = 1000;
