+++ src/main/java/lotto/model/lottoInfo/WinningNumber.java
@@ -1,4 +1,4 @@
-package lotto.model;
+package lotto.model.lottoInfo;
 
 import java.util.List;
